requirements.txt
ultralytics
opencv-python-headless
matplotlib
numpy
Pillow
streamlit
